refactor: repackage exception classes
