+++ src/main/java/core/di/exception/BeanDuplicationException.java
@@ -1,4 +1,4 @@
-package core.di.factory;
+package core.di.exception;
 
 public class BeanDuplicationException extends RuntimeException {
     public BeanDuplicationException(Class<?> type) {
